\name{colinmisc-package}
\alias{colinmisc-package}
\alias{colinmisc}
\docType{package}
\title{
Colin's Miscellaneous Functions 
}
\description{
Miscellaneous functions I use regularly.
}
\details{
\tabular{ll}{
Package: \tab colinmisc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-13\cr
License: \tab GNU GENERAL PUBLIC LICENSE, Version 3 or later, http://www.gnu.org/licenses/gpl-3.0.html \cr
}
}
\author{
Colin Sheppard, Research Engineer, Schatz Energy Research Center, <colin@humboldt.edu>
}
\references{
%%~~ Literature or other references for background information ~~
}
\keyword{ colinmisc }
\seealso{
\code{\link[snow:plotrix]{}}
}
\examples{
}
